requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
langchain-text-splitters
sentence-transformers
chromadb
pypdf
python-dotenv
python-multipart
